Adds appreciate command and modal
This PR adds an /appreciate command so that the user can write a message
of appreciation for their fellow clanmate(s). The command should be
followed by the user(s) they want to appreciate, which will open up a
modal where they can write out their appreciation message.

The modal also has an input asking if they would like their message
to be anonymous or not (taking Yes, No, Y or N as answers)

The message will then be posted to the Appreciation Station channel
by RuneClock, and their message will be added to the appreciate sheet
for maintenance purposes.